Refactor setup readiness config field builders
